requirements.txt
Pillow
FancyLogger
pyopencl
numpy